+++ CHANGELOG.md
@@ -12,7 +12,7 @@ TBD
 #### Enhancements
 
 - Support Xcode 7.3 and Swift 2.2
-- Add `SystemCurlAnimator` to support `SystemCurlFromTop` and `SystemCurlFromBottom` transition animations
+- Add `SystemPageCurlAnimator` to support `SystemPageCurlFromTop` and `SystemPageCurlFromBottom` transition animations
 
 #### Bugfixes